Fix compiler warning

+++ ares/ares_process.c
@@ -158,7 +158,7 @@ static void write_tcp_data(ares_channel channel, fd_set *write_fds, time_t now)
               vec[n].iov_len = sendreq->len;
               n++;
             }
-          wcount = writev(server->tcp_socket, vec, n);
+          wcount = (ssize_t)writev(server->tcp_socket, vec, n);
           free(vec);
           if (wcount < 0)
             {
